+++ packages/liveblocks-react-lexical/.eslintrc.cjs
@@ -67,7 +67,7 @@ module.exports = {
       },
     },
     {
-      files: ["rollup.config.mjs"],
+      files: ["rollup.config.js"],
       env: {
         node: true,
       },