+++ starter-kits/nextjs-starter-kit/app/api/notifications/email/route.ts
@@ -9,7 +9,8 @@ import { addedToDocumentEmail } from "./addedToDocumentEmail";
 import { textMentionEmail } from "./textMentionEmail";
 import { threadEmail } from "./threadEmail";
 
-// TODO add link to guide and more info
+// Learn more
+// https://liveblocks.io/docs/guides/how-to-create-a-notification-settings-panel
 
 // Add your Liveblocks secret key from https://liveblocks.io/dashboard/apiKeys
 const liveblocks = new Liveblocks({
@@ -17,11 +18,15 @@ const liveblocks = new Liveblocks({
 });
 
 // Add your webhook secret key from a project's webhooks dashboard
-const webhookHandler = new WebhookHandler(
-  process.env.LIVEBLOCKS_WEBHOOK_SECRET_KEY as string
-);
+const WEBHOOK_SECRET_KEY = process.env.LIVEBLOCKS_WEBHOOK_SECRET_KEY;
 
 export async function POST(request: Request) {
+  if (!WEBHOOK_SECRET_KEY) {
+    return new Response("To use email notifications, first set up webhooks", {
+      status: 400,
+    });
+  }
+
   const body = await request.json();
   const headers = request.headers;
 
@@ -29,6 +34,7 @@ export async function POST(request: Request) {
 
   try {
     // Verify if this is a real webhook request
+    const webhookHandler = new WebhookHandler(WEBHOOK_SECRET_KEY);
     event = webhookHandler.verifyRequest({
       headers: headers,
       rawBody: JSON.stringify(body),