+++ CHANGELOG.md
@@ -1,5 +1,48 @@
 ## vNEXT (not yet published)
 
+## v2.20.0
+
+### `@liveblocks/client`
+
+- Implement a proxy factory for `UserNotificationSettings` object to return
+  `null` to prevent any errors when accessing a disabled notification channel.
+
+### `@liveblocks/node`
+
+- Implement a proxy factory for `UserNotificationSettings` object to return
+  `null` to prevent any errors when accessing a disabled notification channel.
+
+### `@liveblocks/react`
+
+- Add optional `useRoom({ allowOutsideRoom: true })` option. When this option is
+  set, the hook will return `null` when used outside of a room, whereas the
+  default behavior of the hook is be to throw.
+- Implement a proxy factory for `UserNotificationSettings` object to return
+  `null` to prevent any errors when accessing a disabled notification channel.
+
+### `@liveblocks/react-ui`
+
+- Improve mentions behavior around whitespace, fixing a regression introduced in
+  `v2.18.3` when we added support for whitespace _within_ mentions.
+- Prevent mention suggestions from scrolling instead of flipping when there’s
+  enough space on the other side (e.g. moving from top to bottom).
+- Improve event propagation in the formatting toolbar of `Composer`.
+
+## v2.19.0
+
+### `@liveblocks/*`
+
+- Output ES modules by default (but CJS builds are still included)
+- Modernize internal build tool settings
+
+### `@liveblocks/node`
+
+- Allow passing optional AbortSignal to all client methods
+- Fix bug in encoding of error information in the LiveblocksError when an API
+  call fails (thanks for reporting, @robcresswell!)
+- Fix `getStorageDocument("my-room", "json")` typing in its output `LiveMap`
+  instances as `ReadonlyMap` instead of serialized plain objects.
+
 ## v2.18.3
 
 ### `@liveblocks/node`